requirements.txt
telethon==1.36.0
openai==1.54.3
python-telegram-bot==21.6
APScheduler==3.10.4
python-dotenv==1.0.1

